agent/requirements.txt
langchain>=0.1.0,<0.2.0
langchain-openai>=0.1.0,<0.2.0
langchain-community>=0.0.20,<0.1.0
langchain-core>=0.1.0,<0.2.0
pydantic>=2.0.0
duckduckgo-search>=5.0.0
requests>=2.31.0
python-dotenv>=1.0.0
qrcode>=7.4.2
Pillow>=10.0.0
geopy>=2.4.0
pyTelegramBotAPI>=4.14.0

